job-market/pyproject.toml
[project]
name = "ai-engineer-research"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.77.1",
    "beautifulsoup4>=4.14.3",
    "html5lib>=1.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "lxml>=6.0.2",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[tool.uv]
exclude-newer = "7 days"
